+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Added `alwaysLeafCategoryAuction` settings parameter to run auctions only over leaf categories
+- Added `activateDebugSponsoredTags` settings parameter to display the (Ad) tag next to sponsored products
 
 ## [2.1.0] - 2025-03-11